Verifying trace plugin schema
