+++ Dockerfile.python2
@@ -1,5 +1,7 @@
 FROM lambci/lambda:build-python2.7
 
+RUN pip install -U setuptools
+
 RUN mkdir -p /var/lib/iopipe
 
 WORKDIR /var/lib/iopipe
