+++ Dockerfile.python3
@@ -1,5 +1,7 @@
 FROM lambci/lambda:build-python3.6
 
+RUN pip install -U setuptools
+
 RUN mkdir -p /var/lib/iopipe
 
 WORKDIR /var/lib/iopipe
